Added custom function in, reupdated, refactored and norminette tokenizer.
Still working on tokenizer.c
