+++ Makefile
@@ -56,6 +56,7 @@ SRC_FILES := main.c \
                 tokenizer/tokenizer_utils_1.c \
 				tokenizer/tokenizer_utils_2.c \
 				tokenizer/tokenizer_utils_3.c \
+				tokenizer/tokenizer_utils_4.c \
                 tokenizer/tokenizer.c \
                 parser/join_tokens.c \
 				parser/parse_double_quotes_utils_1.c \
