+++ src/tokenizer/tokenizer_utils_3.c
@@ -1,64 +1,67 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   tokenizer_utils_2.c                                :+:      :+:    :+:   */
+/*   tokenizer.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/19 12:33:25 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/10 12:38:57 by axlee            ###   ########.fr       */
+/*   Created: 2024/04/19 12:16:40 by gyong-si          #+#    #+#             */
+/*   Updated: 2024/05/23 02:08:53 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-// check if character is space
-int	ft_iswhitespace(char *line)
+void	handle_remaining_text(char **line, t_token **token_lst)
 {
-	if (*line == ' ')
-		return (1);
-	return (0);
-}
+	char	*start;
+	char	*text;
 
-int	ft_isbackslash(char *line)
-{
-	if (*line == '\\')
-		return (1);
-	return (0);
-}
-
-int	ft_issemicolon(char *line)
-{
-	if (*line == ';')
-		return (1);
-	return (0);
-}
-
-char	*ft_copy(char *dest, const char *src, size_t n)
-{
-	size_t	i;
-
-	i = 0;
-	while (i < n && src[i] != '\0')
+	while (ft_iswhitespace(*line))
+		(*line)++;
+	if (**line != '\0')
 	{
-		dest[i] = src[i];
-		i++;
+		start = *line;
+		while (**line && !ft_iswhitespace(*line))
+			(*line)++;
+		text = ft_strndup(start, *line - start);
+		token_add_back(token_lst, text, T_IDENTIFIER);
+		free(text);
 	}
-	dest[i] = '\0';
-	return (dest);
 }
 
-char	*ft_strcat(char *dest, const char *src)
+static void	handle_quoted_escape(char **line, t_token **token_lst)
 {
-	char	*ptr;
+	char	*start;
+	int		length;
+	char	*escaped_token;
 
-	ptr = dest;
-	while (*ptr)
-		ptr++;
-	while (*src)
+	start = *line;
+	(*line)++;
+	while (**line && !ft_iswhitespace(*line) && **line != '|' && **line != '<'
+		&& **line != '>')
+		(*line)++;
+	length = *line - start;
+	escaped_token = (char *)malloc(length + 1);
+	if (escaped_token)
 	{
-		*ptr++ = *src++;
+		ft_strncpy(escaped_token, start, length);
+		escaped_token[length] = '\0';
+		token_add_back(token_lst, escaped_token, T_IDENTIFIER);
+		free(escaped_token);
 	}
-	*ptr = '\0';
-	return (dest);
+}
+
+void	handle_backslash(char **line, t_token **token_lst)
+{
+	char	escaped[2];
+
+	escaped[0] = **line;
+	escaped[1] = '\0';
+	(*line)++;
+	if (**line == '\"' || **line == '\'')
+		handle_quoted_escape(line, token_lst);
+	else
+		token_add_back(token_lst, escaped, T_IDENTIFIER);
+	(*line)++;
 }
